+++ examples/use-globals-request/app.py
@@ -35,13 +35,20 @@ _request = ProcessRequest(routes)
 
 def process(client: Client, use_globals: bool):
     # passing "True" to use_globals
-    process_result = _request.process(client, use_globals=use_globals)
+    processed_request = _request.process(client)
 
-    if process_result:
-        response, request_data = process_result
-        log.log_request(response, request_data)
+    if processed_request:
+        request_data = processed_request.request
+        process_type = processed_request.type
+
+        if process_type == 'route':
+            response = processed_request.route.get_route_response(request_data, use_globals)
+        else:
+            response = processed_request.route.get_callback_response(request_data)
 
         http = http_pyparser.response.make_response(response)
+
+        log.log_request(response, request_data)
         client.send_message(http)
         client.destroy()
 
@@ -52,5 +59,4 @@ print('Server running in http://127.0.0.1:2808')
 
 while True:
     client = server.wait_client()
-    # passing "True" to use_globals
     threading.Thread(target=process, args=(client, True)).start()